backend/account/pyproject.toml
[tool.poetry]
name = "auth technostrelka"
version = "0.1.0"
description = "Сервис аккаунтов для олимпиады по backend разработки"
authors = ["ligets <legen2a208@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
poetry = "^1.8.3"
pydantic-settings = "^2.5.2"
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"
alembic = "^1.13.3"
passlib = "^1.7.4"
bcrypt = "4.0.1"
datetime = "^5.5"
python-multipart = "^0.0.12"
aio-pika = "^9.4.3"
pyjwt = "^2.9.0"
cryptography = "^43.0.3"
pydantic = {extras = ["email"], version = "^2.10.3"}
aiobotocore = "^2.15.2"
bs4 = "^0.0.2"
fastapi-cache2 = "^0.2.2"
redis = "^5.2.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.2"
locust = "^2.32.0"

[tool.pytest.ini_options]
pythonpath = [
    ".", "src"
]
asyncio_mode="auto"
asyncio_default_fixture_loop_scope="session"

[tool.locust]
locustfile = "tests/locust/locustfile.py"
host = "http://localhost:8000"
users = 100
spawn-rate = 10

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


